Struct succinct::int_vec::IntVec

pub struct IntVec<N: NonZero + Unsigned, Block: BlockType = usize> {
    // some fields omitted
}

A vector of N-bit unsigned integers.

Block gives the representation type. N must not exceed the number of bits in Block.

Methods

impl<N, Block> IntVec<N, Block> where N: NonZero + Unsigned, Block: PrimInt

fn block_bits() -> usize

The number of bits per block of storage.

fn element_bits() -> usize

The number of bits per elements.

fn is_packed() -> bool

True if elements are packed one per block.

fn is_aligned() -> bool

True if elements are aligned within blocks.

fn new(n_elements: usize) -> Self

Creates a new vector to hold the given number of elements.

Panics

Panics if any of:

block_bits() < N;

n_elements * N doesn’t fit in a u64; or

ceiling(n_elements * N / block_bits()) doesn’t fit in a usize.

fn len(&self) -> usize

Returns the number of elements in the vector.

fn is_empty(&self) -> bool

Is the vector empty?

fn get(&self, element_index: usize) -> Block

Returns the element at the given index.

fn set(&mut self, element_index: usize, element_value: Block)

Sets the element at the given index.

fn get_bit(&self, bit_index: usize) -> bool

Gets the bit at the given position.

fn set_bit(&mut self, bit_index: usize, bit_value: bool)

Sets the bit at the given position.

fn iter(&self) -> Iter<N, Block>

Gets an iterator over the elements of the vector.