Struct succinct::int_vec::IntVecBuilder

pub struct IntVecBuilder<Block: BlockType = usize> {
    // some fields omitted
}

Builder for configuring and constructing an IntVec.

Methods

impl<Block: BlockType> IntVecBuilder<Block>

fn new(element_bits: usize) -> Self

Creates a new IntVecBuilder with element_bits bits per elements.

fn build(&self) -> IntVec<Block>

Builds the specified IntVec.

Panics

Panics if the size conditions of IntVec::<Block>::is_okay_size() are not met.

fn element_bits(&mut self, element_bits: usize) -> &mut Self

Sets the element size to element_bits.

The elements will range from 0 to 2.pow(element_bits) - 1.

fn n_elements(&mut self, n_elements: u64) -> &mut Self

Sets the initial number of elements.

If n_elements() finds that capacity() has been set to a lower value, it adjust capacity() upward.

fn capacity(&mut self, capacity: u64) -> &mut Self

Sets the size of the initial allocation, which may be larger than the initial number of elements.

If capacity() finds that n_elements() has been set to a higher value, it adjust n_elements() downward.

fn zero_fill(&mut self) -> &mut Self

Zero-fill the new vector’s data.

fn block_fill(&mut self, block: Block) -> &mut Self

Fill the vector’s data with the specified block. This will align as a block, which may not align with elements in any particular way. It’s not yet specified how the elements are laid out.

fn element_fill(&mut self, element: Block) -> &mut Self

Fill the vector’s data with the given element.