Struct succinct::int_vec::IntVector

pub struct IntVector<Block: BlockType = usize> {
    // some fields omitted
}

Uncompressed vector of k-bit unsigned integers.

The element width k is determined at vector creation time.

Block gives the representation type. The element width k can never exceed the number of bits in Block.

Methods

impl<Block: BlockType> IntVector<Block>

fn new(element_bits: usize) -> Self

Creates a new integer vector.

Arguments

element_bits — the size of each element in bits; hence elements range from 0 to 2.pow(element_bits) - 1.

Result

The new, empty integer vector.

fn with_capacity(element_bits: usize, capacity: u64) -> Self

Creates a new, empty integer vector, allocating sufficient storage for capacity elements.

fn block_with_capacity(element_bits: usize, block_capacity: usize) -> Self

Creates a new, empty integer vector, allocating block_capacity blocks of storage.

fn with_fill(element_bits: usize, len: u64, value: Block) -> Self

Creates a new integer vector containing len copies of value.

fn block_with_fill(element_bits: usize, block_len: usize, value: Block) -> Self

Creates a new integer vector containing block_len copies of the block value.

The length of the new vector will be the number of elements of size element_bits that fit in block_len blocks.

fn get_random(&self, bit_offset: u64, element_bits: usize, element_index: u64) -> Block

Returns the element at a given index, also given an arbitrary element size and bit offset.

This computes the location of the element_indexth element supposing that elements are element_bits side, then adds bit_offset additional bits and returns the element_bits-bit value found at that location.

Panics

Panics if the referenced bits are out of bounds. Bounds are considered to the end of the support array, even if that goes past the last element of the IntArray.

fn set_random(&mut self, bit_offset: u64, element_bits: usize, element_index: u64, element_value: Block)

Sets the element at a given index to a given value, also given an arbitrary element size and bit offset.

This computes the location of the element_indexth element supposing that elements are element_bits side, then adds bit_offset additional bits and modifies the element_bits-bit value found at that location.

Panics

Panics if the referenced bits are out of bounds. Bounds are considered to the end of the support array, even if that goes past the last element of the IntArray.

Debug mode only: Panics if element_value is too large to fit in the element size. (TODO: What’s the right thing here?)

fn push(&mut self, element_value: Block)

Pushes an element onto the end of the vector, increasing the length by 1.

fn pop(&mut self) -> Option<Block>

Removes and returns the last element of the vector, if present.

fn capacity(&self) -> u64

The number of elements the vector can hold without reallocating.

fn block_capacity(&self) -> usize

The number of blocks of elements the vector can hold without reallocating.

fn resize(&mut self, n_elements: u64, fill: Block)

Resizes to the given number of elements, filling if necessary.

fn block_resize(&mut self, n_blocks: usize, fill: Block)

Resizes to the given number of blocks, filling if necessary.

fn reserve(&mut self, additional: u64)

Reserves capacity for at least additional more elements to be inserted in the given IntVector<Block>.

The collection may reserve more space to avoid frequent reallocations.

Panics

Panics if the size conditions of IntVector::<Block>::is_okay_size() are not met. This will happen if the total number of bits overflows u64.

fn block_reserve(&mut self, additional: usize)

Reserves capacity for at least additional blocks of values to be inserted.

The collection may reserve more space to avoid frequent reallocations.

Panics

Panics if the number of blocks overflows a usize.

fn reserve_exact(&mut self, additional: u64)

Reserves capacity for at least additional more elements to be inserted in the given IntVector<Block>.

Unlike reserve, does nothing if the capacity is already sufficient.

Panics

Panics if the size conditions of IntVector::<Block>::is_okay_size() are not met. This will happen if the total number of bits overflows u64.

fn block_reserve_exact(&mut self, additional: usize)

Reserves capacity for at least additional blocks of values to be inserted.

Unlike reserve_block, does nothing if the capacity is already sufficient.

The collection may reserve more space to avoid frequent reallocations.

Panics

Panics if the number of blocks overflows a usize.

fn shrink_to_fit(&mut self)

Shrinks the capacity to just fit the number of elements.

fn truncate(&mut self, n_elements: u64)

Shrinks to the given size.

Does nothing if n_elements is greater than the current size.

fn block_truncate(&mut self, n_blocks: usize)

Shrinks to the given number of blocks.

Does nothing if n_blocks is greater than the current blocks.

fn clear(&mut self)

Sets the size to 0 while retaining the allocated storage.

fn iter(&self) -> Iter<Block>

Gets an iterator over the elements of the vector.

fn is_block_sized(&self) -> bool

True if the element size matches the block size.

fn is_aligned(&self) -> bool

True if elements are aligned within blocks.