Module succinct::int_vector

Bit-packed vectors of k-bit unsigned integers.

Structs

IV
A vector of k-bit unsigned integers, where k is determined at run time.

Iter
An iterator over the elements of an IV.

Traits

IntVec
Types that can be accessed as immutable arrays of integers of limited width.

IntVecMut
Types that can be accessed as mutable arrays of integers of limited width.