Struct succinct::rank::JacobsonRank

pub struct JacobsonRank<Store> { /* fields omitted */ }

Jacobson’s rank structure for fast rank queries over a BitVec.

Construct with JacobsonRank::new.

Methods

impl<Store: BitVec> JacobsonRank<Store>

pub fn new(bits: Store) -> Self

Creates a new rank support structure for the given bit vector.

Important traits for &'a mut I
impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized,

pub fn inner(&self) -> &Store

Borrows a reference to the underlying bit store.

pub fn into_inner(self) -> Store

Returns the underlying bit store.