Struct succinct::select::BinSearchSelect

pub struct BinSearchSelect<Rank> {
    // some fields omitted
}

Performs a select query by binary searching rank queries.

Methods

impl<Rank: RankSupport> BinSearchSelect<Rank>

Creates a new binary search select support based on a rank support.

fn new(rank_support: Rank) -> Self

Creates a new binary search selection support given a rank support.

fn inner(&self) -> &Rank

Borrows a reference to the underlying rank support.

fn into_inner(self) -> Rank

Returns the underlying rank structure.