Trait succinct::select::SelectSupport

pub trait SelectSupport {
    type Over: Copy;
    fn select(&self, index: u64, value: Self::Over) -> Option<u64>;
}

Supports select queries over associated type Over.

Associated Types

type Over: Copy

The type of value that we can search for.

Required Methods

fn select(&self, index: u64, value: Self::Over) -> Option<u64>

Returns the position of the indexth occurrence of value.