Struct succinct::stream::BitBuffer

pub struct BitBuffer<Inner = BitVector> { /* fields omitted */ }

A bit buffer can be used to read bits from or write bits to an underlying bit vector.

Methods

impl<Block: BlockType> BitBuffer<BitVector<Block>>

pub fn new() -> Self

Creates a new, empty bit buffer.

pub fn with_capacity(capacity: u64) -> Self

Creates a new, empty bit buffer with the given capacity (in bits) preallocated.

impl<Inner: BitVec> BitBuffer<Inner>

pub fn from(input: Inner) -> Self

Creates a new bit buffer for reading from a bit vector.

pub fn append(vec: Inner) -> Self

Creates a new bit buffer for appending to a bit vector.

pub fn seek(&mut self, position: u64) -> Result<()>

Moves the position for the next read or write.

impl<Inner> BitBuffer<Inner>

pub fn into_inner(self) -> Inner

Returns the bit vector underlying the bit buffer.

Important traits for &'a mut I
impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized,

pub fn inner(&self) -> &Inner

Gives access to the bit vector underlying the bit buffer.

pub fn position(&self) -> u64

The position in the bit buffer where the next read or write will occur.