Struct succinct::BV

pub struct BV<Block: BlockType = usize>(_);

A bit vector implementation.

Methods

impl<Block: BlockType> BV<Block>

fn new() -> Self

Creates a new, empty bit vector.

fn with_capacity(capacity: u64) -> Self

Creates a new, empty bit vector with space allocated for capacity bits.

Panics

Panics if capacity is too large. In particular the number of blocks required by the capacity (capacity / Block::nbits()) must fit in a usize.

fn block_with_capacity(capacity: usize) -> Self

Creates a new, empty bit vector with space allocated for capacity blocks.

fn with_fill(len: u64, value: bool) -> Self

Creates a new bit vector of len bits initialized to value.

Panics

Panics if len is too large. In particular the number of blocks required by the capacity (len / Block::nbits()) must fit in a usize.

fn block_with_fill(block_len: usize, value: Block) -> Self

Creates a new bit vector of block_len blocks initialized to value.

fn capacity(&self) -> u64

How many bits the bit vector can hold without reallocating.

fn block_capacity(&self) -> usize

How many blocks the bit vector can hold without reallocating.

fn resize(&mut self, new_len: u64, value: bool)

Resizes the bit vector to the given number of elements, filling if necessary.

Panics

Panics if new_len is too large. In particular the number of blocks required by the capacity (new_len / Block::nbits()) must fit in a usize.

fn block_resize(&mut self, new_len: usize, value: Block)

Resizes the bit vector to the given number of blocks, filling if necessary.

fn reserve(&mut self, additional: u64)

Reserves capacity for at least additional more bits to be inserted.

The collection may reserve more space to avoid frequent reallocations.

Panics

Panics if the number of blocks overflows a usize.

fn block_reserve(&mut self, additional: usize)

Reserves capacity for at least additional blocks of bits to be inserted.

The collection may reserve more space to avoid frequent reallocations.

Panics

Panics if the number of blocks overflows a usize.

fn reserve_exact(&mut self, additional: u64)

Reserves capacity for at least additional more bits to be inserted.

Unlike reserve, does nothing if the capacity is already sufficient.

Panics

Panics if the number of blocks overflows a usize.

fn block_reserve_exact(&mut self, additional: usize)

Reserves capacity for at least additional more blocks of bits to be inserted.

Unlike reserve_block, does nothing if the capacity is already sufficient.

Panics

Panics if the number of blocks overflows a usize.

fn shrink_to_fit(&mut self)

Shrinks the capacity to just fit the number of elements.

fn truncate(&mut self, len: u64)

Shrinks to the given size.

Does nothing if len is greater than the current size.

fn block_truncate(&mut self, block_len: usize)

Shrinks to the given size in blocks.

Does nothing if block_len is greater than the current size in blocks.

fn clear(&mut self)

Sets the size to 0 while retaining the allocated storage.

fn iter(&self) -> Iter<Block>

Returns an iterator over the bits of the bit vector