Struct succinct::BinSearchSelect

pub struct BinSearchSelect<'a, Rank: RankSupport + 'a> {
    // some fields omitted
}

Performs a select query by binary searching rank queries.

Methods

impl<'a, Rank: RankSupport + 'a> BinSearchSelect<'a, Rank>

Creates a new binary search select support based on a rank support.

fn new(rank_support: &'a Rank) -> Self

Creates a new binary search selection support given a rank support.