Module succinct::vector

Bit-packed vectors of k-bit unsigned integers.

Reexports

pub use super::*;

Structs

IntVec
A vector of k-bit unsigned integers, where k is determined at run time.

IntVecBuilder
Builder for configuring and constructing an IntVec.

Iter
An iterator over the elements of an IntVec.

Enums

Fill
Describes how to initialize the memory of an IntVec.

Traits

IntVector
Types that can be accessed as immutable arrays of integers of limited width.

IntVectorMut
Types that can be accessed as mutable arrays of integers of limited width.