Struct succinct::vector::IntVec

pub struct IntVec<Block: BlockType = usize> {
    // some fields omitted
}

A vector of k-bit unsigned integers, where k is determined at run time.

Construct with IntVec::new, or for more control, IntVecBuilder.

Block gives the representation type. The element size k can never exceed the number of bits in Block.

Methods

impl<Block: BlockType> IntVec<Block>

fn new(element_bits: usize) -> Self

Creates a new integer vector.

Arguments

element_bits — the size of each element in bits; hence elements range from 0 to 2.pow(element_bits) - 1.

Result

The new, empty integer vector.

fn is_okay_size(element_bits: usize, n_elements: u64) -> bool

Determines whether we can support a vector with the given element size and number of elements.

We cannot support vectors where:

block_bits() < element_bits;

n_elements * element_bits doesn’t fit in a u64; or

n_elements * element_bits / block_bits() (but with the division rounded up) doesn’t fit in a usize.

where block_bits() is the size of the Block type parameter.

fn get_random(&self, bit_offset: u64, element_bits: usize, element_index: u64) -> Block

Returns the element at a given index, also given an arbitrary element size and bit offset.

This computes the location of the element_indexth element supposing that elements are element_bits side, then adds bit_offset additional bits and returns the element_bits-bit value found at that location.

Panics

Panics if the referenced bits are out of bounds. Bounds are considered to the end of the support array, even if that goes past the last element of the IntArray.

fn set_random(&mut self, bit_offset: u64, element_bits: usize, element_index: u64, element_value: Block)

Sets the element at a given index to a given value, also given an arbitrary element size and bit offset.

This computes the location of the element_indexth element supposing that elements are element_bits side, then adds bit_offset additional bits and modifies the element_bits-bit value found at that location.

Panics

Panics if the referenced bits are out of bounds. Bounds are considered to the end of the support array, even if that goes past the last element of the IntArray.

Debug mode only: Panics if element_value is too large to fit in the element size. (TODO: What’s the right thing here?)

fn push(&mut self, element_value: Block)

Pushes an element onto the end of the vector, increasing the length by 1.

fn pop(&mut self) -> Option<Block>

Removes and returns the last element of the vector, if present.

fn capacity(&self) -> u64

The number of elements the vector can hold without reallocating.

fn resize(&mut self, n_elements: u64, fill: Fill<Block>)

Resizes to the given number of elements, filling if necessary.

fn reserve(&mut self, additional: u64)

Reserves capacity for at least additional more elements to be inserted in the given IntVec<Block>.

The collection may reserve more space to avoid frequent reallocations.

Panics

Panics if the size conditions of IntVec::<Block>::is_okay_size() are not met. This will happen if the total number of bits overflows u64.

fn reserve_exact(&mut self, additional: u64)

Reserves capacity for at least additional more elements to be inserted in the given IntVec<Block>.

Unlike reserve, does nothing if the capacity is already sufficient.

Panics

Panics if the size conditions of IntVec::<Block>::is_okay_size() are not met. This will happen if the total number of bits overflows u64.

fn shrink_to_fit(&mut self)

Shrinks the capacity to just fit the number of elements.

fn truncate(&mut self, n_elements: u64)

Shrinks to the given size.

If n_elements is greater than the current size, does nothing.

fn as_block_slice(&self) -> &[Block]

Returns a reference to the backing slice of blocks.

Note that this does not respect element boundaries, and the layout is not specified.

fn as_mut_block_slice(&mut self) -> &mut [Block]

Returns a mutable reference to the backing slice of blocks.

Note that this does not respect element boundaries, and the layout is not specified.

fn clear(&mut self)

Sets the size to 0 while retaining the allocated storage.

fn iter(&self) -> Iter<Block>

Gets an iterator over the elements of the vector.

fn block_bits() -> usize

The number of bits per block of storage.

fn is_packed(&self) -> bool

True if elements are packed one per block.

fn is_aligned(&self) -> bool

True if elements are aligned within blocks.