Trait succinct::bit_vec::IntoRange

pub trait IntoRange<T> {
    fn into_range(self, start: T, limit: T) -> Range<T>;
}

Range polymorphism support.

The idea is to realize partial ranges by providing start limits to fill in the missing bounds.

Examples

use succinct::bit_vec::IntoRange;

assert_eq!((3..5).into_range(0, 8), 3..5);
assert_eq!(( ..5).into_range(0, 8), 0..5);
assert_eq!((3.. ).into_range(0, 8), 3..8);
assert_eq!(( .. ).into_range(0, 8), 0..8);

Required Methods

fn into_range(self, start: T, limit: T) -> Range<T>

Instantiates a range to a structure by provided bounds where bounds are absent.