Struct succinct::rank::JacobsonRank

pub struct JacobsonRank<Store> {
    // some fields omitted
}

Jacobson’s rank structure for fast rank queries over a BitVec.

Construct with JacobsonRank::new.

Methods

impl<Store: BitVec> JacobsonRank<Store>

fn new(bits: Store) -> Self

Creates a new rank support structure for the given bit vector.

fn inner(&self) -> &Store

Borrows a reference to the underlying bit store.

fn into_inner(self) -> Store

Returns the underlying bit store.