Struct succinct::stream::BitBuffer

pub struct BitBuffer<Inner = BitVector> {
    // some fields omitted
}

A bit buffer can be used to read bits from or write bits to an underlying bit vector.

Methods

impl<Block: BlockType> BitBuffer<BitVector<Block>>

fn new() -> Self

Creates a new, empty bit buffer.

fn with_capacity(capacity: u64) -> Self

Creates a new, empty bit buffer with the given capacity (in bits) preallocated.

impl<Inner: BitVec> BitBuffer<Inner>

fn from(input: Inner) -> Self

Creates a new bit buffer for reading from a bit vector.

fn append(vec: Inner) -> Self

Creates a new bit buffer for appending to a bit vector.

fn seek(&mut self, position: u64) -> Result<()>

Moves the position for the next read or write.

impl<Inner> BitBuffer<Inner>

fn into_inner(self) -> Inner

Returns the bit vector underlying the bit buffer.

fn inner(&self) -> &Inner

Gives access to the bit vector underlying the bit buffer.

fn position(&self) -> u64

The position in the bit buffer where the next read or write will occur.